Fix syntactic errors in Generic_Factory
This design for the Generic Factory now appears to work perfectly.  At
the least, the existing code which used the previous version of the
Generic_Factory (before the pointer type was added to the code) now
compiles perfectly and passes unit tests.  As such, it appears that this
code is fully working, although it is a bit of a sloppy design, in my
opinion.