Clean up Generic_Factor code a bit.
Something strange seems to have happened with the submodule.  It was not
compiling but the error messages actually had the wrong version of the
code there.  Very weird.  This was a minor change -- just reformatted
the code -- to force it to update the code properly in git... hopefully
this will address the problem